Resume page implementation
Added the resumes component.
Added the resume component.
Added the link to the resumes page.
Added the Resume model.
Added the resumes service.
Added the enum utility service.
Added a new pipe to conver a number to an array.
Updated cypress version.
Added a new route + child route for the resumes.
Updated the app module with the new imports.
Added various enum for the resume model.
Fixed minor issues.
Documented the code.